requirements.txt
pyvis
websocket-client